refactoring(checkpoints): extracted logic and functions to useCheckpoints hook
