+++ src/hooks/useCheckpoints.ts
@@ -1,16 +1,18 @@
-import { useCallback, useMemo } from "react";
+import { useCallback, useMemo, useState } from "react";
 import { useAppSelector } from "./useAppSelector";
 import {
   selectIsCheckpointsPopupIsVisible,
   selectIsUndoingCheckpoints,
+  selectLatestCheckpointResult,
   setIsCheckpointsPopupIsVisible,
   setIsUndoingCheckpoints,
   setLatestCheckpointResult,
 } from "../features/Checkpoints/checkpointsSlice";
 import { useAppDispatch } from "./useAppDispatch";
 import { useRestoreCheckpoints } from "./useRestoreCheckpoints";
-import { Checkpoint } from "../features/Checkpoints/types";
+import { Checkpoint, FileChanged } from "../features/Checkpoints/types";
 import { debugRefact } from "../debugConfig";
+import { STUB_RESTORED_CHECKPOINT_DATA } from "../__fixtures__/checkpoints";
 
 export const useCheckpoints = () => {
   const dispatch = useAppDispatch();
@@ -20,27 +22,53 @@ export const useCheckpoints = () => {
   );
   const isUndoingCheckpoints = useAppSelector(selectIsUndoingCheckpoints);
 
+  const latestRestoredCheckpointsResult = useAppSelector(
+    selectLatestCheckpointResult,
+  );
+
+  const [shouldMockBeUsed, setShouldMockBeUsed] = useState(false);
+
+  const { reverted_changes, reverted_to } = shouldMockBeUsed
+    ? STUB_RESTORED_CHECKPOINT_DATA
+    : latestRestoredCheckpointsResult;
+
+  const allChangedFiles = reverted_changes.reduce<
+    (FileChanged & { workspace_folder: string })[]
+  >((acc, change) => {
+    const filesWithWorkspace = change.files_changed.map((file) => ({
+      ...file,
+      workspace_folder: change.workspace_folder,
+    }));
+    return [...acc, ...filesWithWorkspace];
+  }, []);
+
+  const wereFilesChanged = useMemo(() => {
+    return allChangedFiles.length > 0;
+  }, [allChangedFiles]);
+
   const shouldCheckpointsPopupBeShown = useMemo(() => {
-    // TODO: show if isVisible and files data is not empty
     return isCheckpointsPopupVisible && !isUndoingCheckpoints;
   }, [isCheckpointsPopupVisible, isUndoingCheckpoints]);
 
-  const handleUndo = useCallback(
-    async (checkpoints: Checkpoint[]) => {
-      await restoreChangesFromCheckpoints(checkpoints);
-      dispatch(setIsUndoingCheckpoints(true));
-    },
-    [dispatch, restoreChangesFromCheckpoints],
-  );
+  const handleUndo = useCallback(async () => {
+    await restoreChangesFromCheckpoints(
+      latestRestoredCheckpointsResult.checkpoints_for_undo,
+    );
+    dispatch(setIsUndoingCheckpoints(true));
+  }, [
+    dispatch,
+    restoreChangesFromCheckpoints,
+    latestRestoredCheckpointsResult.checkpoints_for_undo,
+  ]);
 
   const handleRestore = useCallback(
     async (checkpoints: Checkpoint[] | null) => {
       if (!checkpoints) return;
-      dispatch(setIsUndoingCheckpoints(false));
       const restoredChanges =
         await restoreChangesFromCheckpoints(checkpoints).unwrap();
       debugRefact(`[DEBUG]: restoredChanges received: `, restoredChanges);
       const actions = [
+        dispatch(setIsUndoingCheckpoints(false)),
         setLatestCheckpointResult(restoredChanges),
         setIsCheckpointsPopupIsVisible(true),
       ];
@@ -49,10 +77,29 @@ export const useCheckpoints = () => {
     [dispatch, restoreChangesFromCheckpoints],
   );
 
+  const handleFix = useCallback(() => {
+    dispatch(setIsCheckpointsPopupIsVisible(false));
+  }, [dispatch]);
+
+  // TODO: remove when fully tested
+  const handleShouldMockBeUsedChange = useCallback(() => {
+    setShouldMockBeUsed((prev) => !prev);
+  }, []);
+  // end TODO
+
   return {
     shouldCheckpointsPopupBeShown,
     handleUndo,
     handleRestore,
+    handleFix,
+    // TODO: remove when fully tested
+    handleShouldMockBeUsedChange,
+    shouldMockBeUsed,
+    // end TODO
     isLoading,
+    reverted_changes,
+    reverted_to,
+    wereFilesChanged,
+    allChangedFiles,
   };
 };
